17강 완료

+++ Assets/02_Scripts/CoinScript.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 //04_Prefabs-Coin
 public class CoinScript : MonoBehaviour {
     public float speed = 1.3f; //코인의 속도-배경의 속도와 똑같이
-    public float coinSize = 1;
+    public double coinSize = 1;
     Transform playerTr;
     private void Start() {
         playerTr = GameObject.FindWithTag("Player").transform;
